backend-session-manager/requirements.txt
fastapi>=0.115,<1.0
httpx>=0.27,<1.0
pydantic-settings>=2.6,<3.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
pytest>=8.3,<9.0
sqlalchemy>=2.0,<3.0
uvicorn>=0.32,<1.0
